requirements.txt
requests==2.31.0
terminaltables==3.1.10
python-dotenv==0.9.1
